// ProjectMomoDonation.FrontEnd/src/app/feature/user/model/program-create.model.ts
export interface ProgramCreateModel {

    title: string;
    avatarUrl: string;
    urlSlug: string;
    shortTitle: string;
    story: string;
    dateStart: string;
    dateEnd: string;
    totalDonate : number;
    donationGoal: number;
    categoryName: string;
    organizationName: string;
  }
  